requirements.txt
--index-url https://download.pytorch.org/whl/cu128
torch
torchvision
torchaudio

transformers
datasets
scikit-learn
pyyaml
numpy
pandas
pyarrow
